backend/pyproject.toml
[tool.poetry]
name = "genealogy-backend"
version = "0.1.0"
description = "GEDCOM X Python Fast API Backend"
authors = ["Simon Bragg <36492124+maiorsi@users.noreply.github.com>"]
license = "MIT"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.93.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
beanie = "^1.17.0"
pydantic = {extras = ["email"], version = "^1.10.5"}
gedcomx-python = "^0.1.6"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.15.0"
black = "^22.3.0"
ipython = "^8.3.0"
ipdb = "^0.13.9"
rstcheck = "^5.0.0"
Sphinx = "^4.5.0"
pyupgrade = "^2.32.1"
autoflake = "^1.4"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"

[tool.isort]
profile = "black"
src_paths = ["gedcomx", "tests"]

[tool.poetry-version-plugin]
source = "init"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
